rmfstreamer corefiles are failed to upload
Reason for change: Bootup upload was done for minidump/coredumps
                   If minidumps comes first it is cleaning the bootup
                   check flags and because of this coredumps clean script
                   is cleaning the previous dumps which is not uploaded
Test Procedure: Ensure coredump upload/Minidump Upload
Risks: Low

+++ uploadDumpsHDD.sh
@@ -187,7 +187,7 @@ deleteAllButTheMostRecentFile()
 
 cleanup()
 {
-    if [ -z "$WORKING_DIR" ] ; then
+    if [ -z "$WORKING_DIR" ] || [ -z "$(ls -A $WORKING_DIR 2> /dev/null)" ] ; then
         logMessage "WORKING_DIR is empty!!!"
         return
     fi
@@ -227,7 +227,9 @@ cleanup()
             touch "$ON_STARTUP_DUMPS_CLEANED_UP"
        fi
     else
-       rm -rf /opt/.upload_on_startup
+       if [ $DUMP_FLAG -eq 1 ];then
+            rm -rf /opt/.upload_on_startup
+       fi
     fi
 }
